Changed WidgetGeometry variables type from int to float

+++ src/Widgets/WidgetGeometry.hpp
@@ -4,10 +4,10 @@
 
 struct WidgetGeometry
 {
-    int x { 0 };
-    int y { 0 };
-    int width { 0 };
-    int height { 0 };
+    float x { 0.0 };
+    float y { 0.0 };
+    float width { 0.0 };
+    float height { 0.0 };
 
     bool operator==(const WidgetGeometry& geometry) const
     {
